replayer: Always run on the empty input
This mimics libFuzzer's behavior and also ensures the fuzz target is
executed at least once even if no inputs are provided.

Also remove the "running on N inputs" line from the logs: It is hard to
get the count to be accurate now that directory inputs are supported and
the empty input is always tested. We could think about brining it back
in the form of a "Finished running N inputs" line printed at the end.